+++ .github/release-notes.md
@@ -1,1 +1,2 @@
-* Fixed demos not being detected properly
+* Fixed error when launching in setup mode
+* Added title to dialog when picking directory